+++ openeogeotrellis/backend.py
@@ -1898,7 +1898,7 @@ class GpsBatchJobs(backend.BatchJobs):
         if len(udf_runtimes) == 1:
             udf_runtime = udf_runtimes.pop()
             if udf_runtime is not None and "image-name" not in job_options and udf_runtime[1] is not None:
-                job_options["image-name"] = udf_runtime[0] + udf_runtime[1].replace(".","")
+                job_options["image-name"] = udf_runtime[0].lower() + udf_runtime[1].replace(".","")
         elif len(udf_runtimes) > 1:
             log.warning(f"Multiple UDF runtimes detected in the process graph: {udf_runtimes}. Running with default environment.")