+++ openeogeotrellis/config/config.py
@@ -298,7 +298,5 @@ class GpsBackendConfig(OpenEoBackendConfig):
     """
     s3_region_proxy_endpoints: Dict[str, str] = attrs.Factory(dict)
 
-    """
-    The group ID for the batch job results directory (in non-kubernetes deployments).
-    """
-    non_kube_batch_job_results_dir_group: str = "openeo_results"
+    # FreeIPA server to use (for user lookup/creation)
+    freeipa_server: Optional[str] = os.environ.get("OPENEO_FREEIPA_SERVER", None)